# backend/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "description": "API Documentation",
      "version": "1.0.0",
      "title": "Video Games Search Engine",
      "contact": {
        "email": "example@example.com"
      },
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "servers": [
      {
        "url": "/"
      }
    ],
    "tags": [
      {
        "name": "Video Games Request",
        "description": "Example API for searching for books"
      }
    ],
    "paths": {
      "/api/getTags": {
        "get": {
          "tags": [
            "Get distinct tags for filtering"
          ],
          "summary": "Return filters",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
     },
      "/api/getTest": {
        "get": {
          "tags": [
            "Test Tags"
          ],
          "summary": "Test to check connection",
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
     },
     "/api/search?size={size}&from={from}&developer={developer}&name={name}&platforms={platforms}&genres={genres}&publisher={publisher}&categories={categories}": {
        "parameters": [
          {
            "name": "size",
            "in": "path",
            "required": false,
            "description": "results to show per page",
            "type": "string"
          },
          {
            "name": "from",
            "in": "path",
            "required": false,
            "description": "starting Point",
            "type": "string"
          },
          {
            "name": "name",
            "in": "path",
            "required": false,
            "description": "Name of game",
            "type": "string"
          },
          {
            "name": "platforms",
            "in": "path",
            "required": false,
            "description": "Game platform",
            "type": "string"
          },
          {
            "name": "developer",
            "in": "path",
            "required": false,
            "description": "Game developer",
            "type": "string"
          },
          {
            "name": "genres",
            "in": "path",
            "required": false,
            "description": "Game genres",
            "type": "string"
          },
          {
            "name": "publisher",
            "in": "path",
            "required": false,
            "description": "Publisher of the game",
            "type": "string"
          },
          {
            "name": "categories",
            "in": "path",
            "required": false,
            "description": "Game categories",
            "type": "string"
          }
        ],
        "get": {
          "tags": [
            "Search"
          ],
          "summary": "Search for video games",
          "responses": {
            "200": {
              "description": "OK"
              }
            }  
          }
        },
        "/api/search?size={size}&name={name}&from={from}": {
          "parameters": [
            {
              "name": "name",
              "in": "path",
              "required": false,
              "description": "Name of game",
              "type": "string"
            },
            {
              "name": "from",
              "in": "path",
              "required": false,
              "description": "from",
              "type": "string"
            },
            {
              "name": "size",
              "in": "path",
              "required": false,
              "description": "results to show per page",
              "type": "string"
            }
          ],
          "get": {
            "tags": [
              "Search"
            ],
            "summary": "Search for video games",
            "responses": {
              "200": {
                "description": "OK"
                }
              }  
            }
        },
        "/api/search?name={name}&developer={developer}&platforms={platforms}": {
          "parameters": [
            {
              "name": "name",
              "in": "path",
              "required": false,
              "description": "Name of game",
              "type": "string"
            },
            {
              "name": "platforms",
              "in": "path",
              "required": false,
              "description": "Game platform",
              "type": "string"
            },
            {
              "name": "developer",
              "in": "path",
              "required": false,
              "description": "Game developer",
              "type": "date"
            }
          ],
          "get": {
            "tags": [
              "Search"
            ],
            "summary": "Search for video games",
            "responses": {
              "200": {
                "description": "OK"
                }
              }  
            }
          },
          "/api/search": {
            "get": {
              "tags": [
                "Search"
              ],
              "summary": "Search for video games",
              "responses": {
                "200": {
                  "description": "OK"
                  }
                }  
              }
         },
         "/api/search?platforms={platforms}&genres={genres}&categories={categories}": {
          "parameters": [
            {
              "name": "platforms",
              "in": "path",
              "required": false,
              "description": "Game platform",
              "type": "string"
            },
            {
              "name": "genres",
              "in": "path",
              "required": false,
              "description": "Game genres",
              "type": "string"
            },
            {
              "name": "categories",
              "in": "path",
              "required": false,
              "description": "Game categories",
              "type": "string"
            }
          ],
          "get": {
            "tags": [
              "Search"
            ],
            "summary": "Search for video games",
            "responses": {
              "200": {
                "description": "OK"
                }
              }  
            }
          },
          "/api/all": {
           "get": {
              "tags": [
                "Search"
              ],
              "summary": "Search for video games",
              "responses": {
                "200": {
                  "description": "OK"
                  }
                }  
              }
          },
          "/api/search?name={name}": {
            "parameters": [
              {
                "name": "name",
                "in": "path",
                "required": false,
                "description": "name",
                "type": "string"
             }
            ],
            "get": {
              "tags": [
                "Search"
              ],
              "summary": "Search for video games",
              "responses": {
                "200": {
                  "description": "OK"
                  }
                }  
              }
         },
         "/api/search?name={name}?developer={developer}": {
          "parameters": [
            {
              "name": "name",
              "in": "path",
              "required": false,
              "description": "name",
              "type": "string"
           },
           {
            "name": "developer",
            "in": "path",
            "required": false,
            "description": "Game developer",
            "type": "string"
          }
          ],
          "get": {
            "tags": [
              "Search"
            ],
            "summary": "Search for video games",
            "responses": {
              "200": {
                "description": "OK"
                }
              }  
            }
       },
         "/api/suggestor?name={name}": {
          "parameters": [
            {
              "name": "name",
              "in": "path",
              "required": false,
              "description": "name",
              "type": "string"
           }
          ],
          "get": {
            "tags": [
              "Suggestor"
            ],
            "summary": "Search for video games",
            "responses": {
              "200": {
                "description": "OK"
                }
              }  
            }
       }

    }
 }